+++ contracts/compliance/modular/modules/SupplyLimitModule.sol
@@ -67,7 +67,6 @@ import "../../../token/IToken.sol";
 import "./AbstractModule.sol";
 
 contract SupplyLimitModule is AbstractModule {
-
     /// supply limits array
     mapping(address => uint256) private _supplyLimits;
 
@@ -133,5 +132,12 @@ contract SupplyLimitModule is AbstractModule {
     */
     function getSupplyLimit(address _compliance) external view returns (uint256) {
         return _supplyLimits[_compliance];
+    }
+
+    /**
+     *  @dev See {IModule-name}.
+     */
+    function name() public pure returns (string memory _name) {
+        return "SupplyLimitModule";
     }
 }