+++ CHANGELOG.md
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file.
 - Implement a new compliance module `Time Transfers limits` that prevents holders from transfering more token than a specified limit in a given time frame.
 - Implement a new compliance module `Max Balance Module` that prevents an individual holder to own more than a given percentage of the supply.
 - Implement two new compliance modules `Time Exchange limits` and `Monthly Exchange limits` that limit exchanges transfers. These are used to authorized specific trusted exchanges to hold tokens but limited to a certain amount transfered for a given time frame.
+- Add `function name() external pure returns (string memory _name);` to `IModule`. Compliance modules now require a
+  `function name() public pure returns (string memory _name) {
+    return "CountryRestrictModule";
+  }`
+  constant variable to be declared.
 
 ## [4.0.1]